requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
scipy==1.11.4
pyyaml==6.0.1
python-dotenv==1.0.1
joblib==1.4.2
requests==2.31.0
matplotlib==3.8.4
lightgbm==4.3.0
xgboost==2.0.3
optuna==3.6.1
mlflow==2.12.1
imbalanced-learn==0.12.3
shap==0.45.0
networkx==3.2.1
pyspark==3.5.1
delta-spark==3.1.0
pyarrow==15.0.2
aequitas==0.42.0
evidently==0.4.21
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.1
web3==6.11.3
